move offers-found-info and pagination to common ui
