+++ frontend/app/ui/offers-found-info.tsx
